chore: move common cycle test mocks to utils
